feat: finalizar regras de negócio e validações de Autor
- Valida duplicidade e consistência de dados
- Refatora método de salvar/atualizar
- Integra validações com retornos padronizados de erro

+++ src/main/java/com/example/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package com.example.libraryapi.service;
 
+import com.example.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import com.example.libraryapi.model.Autor;
 import com.example.libraryapi.repository.AutorRepository;
+import com.example.libraryapi.repository.LivroReposirory;
+import com.example.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,17 +15,25 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository autorRepository;
+    private final AutorValidator autorValidator;
+    private final LivroReposirory livroReposirory;
 
-    public AutorService(AutorRepository autorRepository) {
+    public AutorService(AutorRepository autorRepository, AutorValidator autorValidator, LivroReposirory livroReposirory) {
         this.autorRepository = autorRepository;
+        this.autorValidator = autorValidator;
+        this.livroReposirory = livroReposirory;
     }
 
-    public Autor salvarAutor(Autor autor) { return autorRepository.save(autor); }
+    public Autor salvarAutor(Autor autor) {
+        autorValidator.validar(autor);
+        return autorRepository.save(autor);
+    }
 
     public void atualizarAutor(Autor autor) {
         if (autor.getId() == null) {
             throw new IllegalArgumentException("Para atualizar é necessario que o autor ja esteja salvo na base");
         }
+        autorValidator.validar(autor);
         autorRepository.save(autor);
     }
 
@@ -31,6 +42,9 @@ public class AutorService {
     }
 
     public void deletarAutor(Autor autor) {
+        if (possuiLivro(autor)){
+            throw new OperacaoNaoPermitidaException("Não é permitido excluir autor que possui livros cadastrados!");
+        }
         autorRepository.delete(autor);
     }
 
@@ -48,5 +62,9 @@ public class AutorService {
         }
 
         return autorRepository.findAll();
+    }
+
+    public boolean possuiLivro(Autor autor) {
+        return livroReposirory.existsByAutor(autor);
     }
 }
